#15.1 Unity RPG Tutorial - Equipping armor

+++ Assets/Scripts/UI/Hand.cs
@@ -55,8 +55,12 @@ public class Hand : MonoBehaviour
         // L'image a la position du curseur de la souris
         icon.transform.position = Input.mousePosition + offset;
 
-        // [DEBUG] : Supprime un item
-        DeleteItem();
+        // Clic gauche et que l'on ne pointe pas sur un élément de l'interface (par exemple un bouton d'action) et qu'on déplace un item
+        if (Input.GetMouseButtonDown(0) & !EventSystem.current.IsPointerOverGameObject() && MyInstance.MyMoveable != null)
+        {
+            // [DEBUG] : Supprime un item
+            DeleteItem();
+        }
     }
 
     /// <summary>
@@ -109,24 +113,20 @@ public class Hand : MonoBehaviour
     /// <summary>
     /// Supprime un item
     /// </summary>
-    private void DeleteItem()
+    public void DeleteItem()
     {
-        // Clic gauche et que l'on ne pointe pas sur un élément de l'interface (par exemple un bouton d'action) et qu'on déplace un item
-        if (Input.GetMouseButtonDown(0) & !EventSystem.current.IsPointerOverGameObject() && MyInstance.MyMoveable != null)
+        // Si c'est un item et qu'il vient d'un emplacement
+        if (MyMoveable is Item && InventoryScript.MyInstance.MyFromSlot != null)
         {
-            // Si c'est un item et qu'il vient d'un emplacement
-            if (MyMoveable is Item && InventoryScript.MyInstance.MyFromSlot != null)
-            {
-                // Vide l'emplacement
-                (MyMoveable as Item).MySlot.Clear();
-            }
-
-            // Libère l'item
-            Drop();
-
-            // Réinitialisation de l'emplacement
-            InventoryScript.MyInstance.MyFromSlot = null;
+            // Vide l'emplacement
+            (MyMoveable as Item).MySlot.Clear();
         }
+
+        // Libère l'item
+        Drop();
+
+        // Réinitialisation de l'emplacement
+        InventoryScript.MyInstance.MyFromSlot = null;
     }
 
 }
